Basic Class Creation
Updated gitignore to ignore unnecessary project files. created an empty text file for data storage. created a basic item object with setter and getter functions. created an item lest to store a collection of items in a hashmap with file reading capabilities (unspecified for this project atm). created a main tester file to test the project as I go.