Update files
Created a toString function for the Item class and rearranged the order of some functions and variables (solely for organization). Did some small updates to the readFile function, but it's still not personalized yet. Also wrote a sample product in the data file to prepare for testing the readFile function and the Item class.